requirements.txt
shapely
geopy
pymodes
requests
numpy
pyrtlsdr
kafka-python